pyproject.toml
[tool.poetry]
name = "nutrimais"
version = "0.1.0"
description = ""
authors = ["Amospk2 <Googletraduzpramin@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.14"}
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9.10"
pyjwt = "^2.10.1"
passlib = {extras = ["bcrypt==4.0.1"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
